demo_repo/requirements.inferred.txt
# Generated by betteruv
# Review before committing for production use.

PyYAML
beautifulsoup4
httpx
pytest
python-dateutil
python-dotenv
requests
sampleapp
